Implement postgresql storage for IP addresses
We now store IP addresses in a relational database.  We're using
postgresql because of its inet data types, although we're not really
using them to their full potential yet.  They'll give us room to grow
in terms of more sophisticated allocation schemes.

For the time being we're just storing individual IP addresses in a
table that's mostly a clone of Netbox's IP address table.  The web
service only implements the minimum behavior to support the
Netbox-compatible controller.

We're using the golang "migration" tool to migrate the db, and to seed
it we've got an ad-hoc sql script.

https://github.com/golang-migrate/migrate