Add an Envoy control plane
At the moment it's a hard-coded proxy config similar to the one at
https://github.com/envoyproxy/envoy/blob/f9773269566002fb79593acc2d7089c897d67ca1/configs/google_com_proxy.v2.yaml
which is used by the Envoy quick-start intro.  It simply proxies
requests on port 10000 to www.acnodal.io.

The config is built in internal/envoy/resource.go so that's what we'll
need to enhance to implement dynamic endpoints.  We'll also need to
wire the EGW web service to the control plane snapshot cache so we can
tell the control plane about updates to the configuration.