First try at a CRD-driven IPAM
We're no longer trying to be Netbox-compatible so we no longer
implement the Netbox routes. Addresses are assigned implicitly when
PureLB announces a new LoadBalancer to the EGW, which it can now do in
the EGW pool manager.

The allocator code is cribbed from PureLB's "local" pool
implementation.

The controllers are copied from the egw-resource-model project.