Add a protocol field to port definitions
We need to be able to specify UDP ports as well as TCP, and the native
k8s ServicePort and EndpointPort objects already have a Protocol
field. It actually makes our code a little cleaner since we can use
the k8s objects so we don't need to define our own Port struct.