Use Patch() instead of Update() to add EPs to LBs
When you create a service using PureLB the agent on each node that has
an endpoint will fire at more or less the same time and announce the
endpoint. This was causing errors because the web service was trying
to add an endpoint using Update() but another thread got there first
so the LB CR was stale.

Now we use the most user-hostile API that I've used in years: Patch().
It doesn't have the concurrency problems that Update() has but getting
it to work is a right royal pain in the ass.