Add retries to GETs to work around eventual consistency bug
When the system is busy we can create an object in a POST request but
fail to read the object on the subsequent GET (i.e., after the 302
redirect).

Now we retry a few times on a GET. This causes legit "not found"
requests to take a while to respond, but it makes operation of the
system more robust.