Use a label to tie endpoints to their owning LB
We were copying the owning LB name from the Spec to a Label in the
endpoint webhook but we were seeing duplicate endpoints so there might
be a race condition in that process. By setting the label here we
eliminate that race.

As part of this commit we're also changing the endpoint's name so that
it includes the owning LB's name. This should be helpful in debugging.