Make lb and rep names more readable
They need to be at least a little bit random to avoid name collisions,
but they can be indicative.

Example LB name: adhoc-cli-00d29f9a2414d608

"adhoc-cli" was the name in the request.

Example endpoint name: 172-30-254-22-8888-tcp-ec7c4a04aeb35d5f

The goal is to make the name useful, so in this case the endpoint
address is 172.30.254.22 and the port is 8888 and the protocol is tcp.