Implement per-LB sharing (or not)
When the LB is created we either add some randomness to its name (if
it's not shareable) or leave the name alone (if it's shareable). If
it's not shareable and another PureLB service with the same name tries
to create an LB then the names won't collide. If it is shareable then
when this happens the PureLB announcer will be redirected to the
existing LB.