Remove address allocator
Allocating addresses in this component didn't work out in the
end. There are too many cases where changes can happen to the
LoadBalancer CR that this component doesn't track. It makes better
sense to move allocation to the resource model operator since it
already has the webhook and controller infrastructure.