Fix unwanted service sharing
Because we used only the service name to build the LB CR name, we
could end up with two services sharing an LB when they shouldn't. Two
services in the same account but belonging to different groups should
create two LBs, but instead they were sharing the same one because the
group name didn't influence the sharing decision.

The group name is a sorta-kinda "namespace" because two services with
the same name in the same group should be shared but two services with
the same name in different groups should be independent. Since the
loadbalancer CRs for both services live in the same k8s namespace, to
make this work we now prepend the group name to the service name. This
way two services with the same name won't collide if they belong to
different groups.