Revert "Remove unused OwningClusterLabel"
This reverts commit 55d0771e6d35ec044f160d6f9b230b76c923ec17.

This has been a learning process. It's taken a while to get the
cluster logic correct so it blocks LB deletion when we want the LB to
not be deleted, but doesn't block it when we want it to be
deleted. We've tried a few different approaches.

We removed the epicv1.OwningClusterLabel label because it could be
longer than k8s allows for a label value, but it turns out that we
need it. In the mean time we've figured out that we can use the client
cluster service's UID as a cluster ID and that's a fixed length so we
won't get k8s errors anymore. So now we want to set that as a label
value like we used to.